JAMES-1692 extract SetMessagesMethod.processDestroy to a class
