+++ server/protocols/jmap/src/main/java/org/apache/james/jmap/methods/SetMessagesMethod.java
@@ -19,51 +19,30 @@
 
 package org.apache.james.jmap.methods;
 
-import java.util.*;
-import java.util.function.Consumer;
+import java.util.Set;
 import java.util.stream.Stream;
+
 import javax.inject.Inject;
 
-import org.apache.james.jmap.exceptions.MessageNotFoundException;
 import org.apache.james.jmap.model.ClientId;
-import org.apache.james.jmap.model.MessageId;
-import org.apache.james.jmap.model.SetError;
 import org.apache.james.jmap.model.SetMessagesRequest;
 import org.apache.james.jmap.model.SetMessagesResponse;
 import org.apache.james.mailbox.MailboxSession;
 import org.apache.james.mailbox.exception.MailboxException;
-import org.apache.james.mailbox.model.MessageRange;
-import org.apache.james.mailbox.store.MailboxSessionMapperFactory;
-import org.apache.james.mailbox.store.mail.MailboxMapperFactory;
-import org.apache.james.mailbox.store.mail.MessageMapper;
-import org.apache.james.mailbox.store.mail.MessageMapper.FetchType;
-import org.apache.james.mailbox.store.mail.model.Mailbox;
 import org.apache.james.mailbox.store.mail.model.MailboxId;
-import org.apache.james.mailbox.store.mail.model.MailboxMessage;
 
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Preconditions;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
 public class SetMessagesMethod<Id extends MailboxId> implements Method {
 
-    private static final int LIMIT_BY_ONE = 1;
-    private static final Logger LOGGER = LoggerFactory.getLogger(SetMessagesMethod.class);
     private static final Method.Request.Name METHOD_NAME = Method.Request.name("setMessages");
     private static final Method.Response.Name RESPONSE_NAME = Method.Response.name("messagesSet");
 
-    private final MailboxMapperFactory<Id> mailboxMapperFactory;
-    private final MailboxSessionMapperFactory<Id> mailboxSessionMapperFactory;
-    private final List<SetMessagesProcessor<Id>> messagesProcessors;
+    private final Set<SetMessagesProcessor<Id>> messagesProcessors;
 
     @Inject
-    @VisibleForTesting SetMessagesMethod(MailboxMapperFactory<Id> mailboxMapperFactory,
-                                         MailboxSessionMapperFactory<Id> mailboxSessionMapperFactory,
-                                         List<SetMessagesProcessor<Id>> messagesProcessors) {
-        this.mailboxMapperFactory = mailboxMapperFactory;
-        this.mailboxSessionMapperFactory = mailboxSessionMapperFactory;
+    @VisibleForTesting SetMessagesMethod(Set<SetMessagesProcessor<Id>> messagesProcessors) {
         this.messagesProcessors = messagesProcessors;
     }
 
@@ -95,60 +74,12 @@ public class SetMessagesMethod<Id extends MailboxId> implements Method {
     }
 
     private SetMessagesResponse setMessagesResponse(SetMessagesRequest request, MailboxSession mailboxSession) throws MailboxException {
-        SetMessagesResponse.Builder responseBuilder = SetMessagesResponse.builder();
-        processDestroy(request.getDestroy(), mailboxSession, responseBuilder);
         return messagesProcessors.stream()
                 .map(processor -> processor.process(request, mailboxSession))
-                .reduce(responseBuilder,
+                .reduce(SetMessagesResponse.builder(),
                         (builder, resp) -> resp.mergeInto(builder) ,
                         (builder1, builder2) -> builder2.build().mergeInto(builder1)
                 )
                 .build();
-    }
-
-    private void processDestroy(List<MessageId> messageIds, MailboxSession mailboxSession, SetMessagesResponse.Builder responseBuilder) throws MailboxException {
-        MessageMapper<Id> messageMapper = mailboxSessionMapperFactory.createMessageMapper(mailboxSession);
-        Consumer<? super MessageId> delete = delete(messageMapper, mailboxSession, responseBuilder);
-
-        messageIds.stream()
-            .forEach(delete);
-    }
-
-    private Consumer<? super MessageId> delete(MessageMapper<Id> messageMapper, MailboxSession mailboxSession, SetMessagesResponse.Builder responseBuilder) {
-        return (messageId) -> {
-            try {
-                Mailbox<Id> mailbox = mailboxMapperFactory
-                        .getMailboxMapper(mailboxSession)
-                        .findMailboxByPath(messageId.getMailboxPath(mailboxSession));
-
-                MailboxMessage<Id> mailboxMessage = getMailboxMessage(messageMapper, messageId, mailbox);
-
-                messageMapper.delete(mailbox, mailboxMessage);
-                responseBuilder.destroyed(messageId);
-            } catch (MessageNotFoundException e) {
-                responseBuilder.notDestroyed(messageId,
-                        SetError.builder()
-                        .type("notFound")
-                        .description("The message " + messageId.serialize() + " can't be found")
-                        .build());
-            } catch (MailboxException e) {
-                LOGGER.error("An error occurred when deleting a message", e);
-                responseBuilder.notDestroyed(messageId,
-                        SetError.builder()
-                        .type("anErrorOccurred")
-                        .description("An error occurred while deleting message " + messageId.serialize())
-                        .build());
-            }
-        };
-    }
-
-    private MailboxMessage<Id> getMailboxMessage(MessageMapper<Id> messageMapper, MessageId messageId, Mailbox<Id> mailbox) 
-            throws MailboxException, MessageNotFoundException {
-
-        Iterator<MailboxMessage<Id>> mailboxMessage = messageMapper.findInMailbox(mailbox, MessageRange.one(messageId.getUid()), FetchType.Metadata, LIMIT_BY_ONE);
-        if (!mailboxMessage.hasNext()) {
-            throw new MessageNotFoundException();
-        }
-        return mailboxMessage.next();
     }
 }
